requirements.txt
pymupdf>=1.24.0
pytesseract>=0.3.10
pillow>=9.0.0
